refactor: 회원과 바우처 콘솔 메시지 분리, 에러 메시지 추가
