+++ spring-basic/src/main/java/org/prgrms/springbasic/utils/enumm/message/ErrorMessage.java
@@ -1,16 +1,22 @@
-package org.prgrms.springbasic.utils.enumm;
+package org.prgrms.springbasic.utils.enumm.message;
 
 public enum ErrorMessage {
 
     PARSING_ERROR("Type the correct number type."),
+
     COMMAND_ERROR("Please check your command."),
+
     NO_VOUCHER("Can't find any voucher."),
+
     NO_CUSTOMER("Can't find any customer."),
-    AFTER_DISCOUNT_PRICE_ERR( """
-            Check the discount information you entered.
-            (0 <= AfterDiscount < BeforeDiscount).
-                            """),
-    NOT_EXIST_ENUM_TYPE("Can't find any enum type");
+
+    DISCOUNT_INFO_ERR( "Check the discount information you entered."),
+
+    NOT_EXIST_ENUM_TYPE("Can't find any enum type"),
+
+    NOT_INSERTED("Nothing was inserted."),
+
+    NOT_UPDATED("Nothing was updated.");
 
     private final String message;
