+++ spring-basic/src/main/java/org/prgrms/springbasic/utils/enumm/path/FilePath.java
@@ -1,4 +1,4 @@
-package org.prgrms.springbasic.utils.enumm;
+package org.prgrms.springbasic.utils.enumm.path;
 
 public enum FilePath {
